place handleResult in util function
