+++ src/pages/api/broker/get-orders.ts
@@ -1,6 +1,6 @@
 import type { NextApiRequest, NextApiResponse } from "next";
-import fetch, { BodyInit, Response } from "node-fetch";
-import { convertResult } from "../../../util/convertResult";
+import fetch, { Response } from "node-fetch";
+import { handleResult } from "../../../util";
 
 type Data = {
   orders: Response;
@@ -11,17 +11,6 @@ const buff = Buffer.from(
   "utf-8",
 );
 const base64EncodedKeys = buff.toString("base64");
-
-const handleResult = async (result: Response) => {
-  try {
-    const data = await convertResult(result);
-    const reason = { error: data };
-    return result.ok ? Promise.resolve(data) : Promise.reject(reason);
-  } catch (error) {
-    const reason = { error };
-    return Promise.reject(reason);
-  }
-};
 
 const account_id = "b75acdbc-3fb6-3fb3-b253-b0bf7d86b8bb"; // public info
 const baseUrl = "https://broker-api.sandbox.alpaca.markets/v1";
